requirements.txt
absl-py==2.2.0
accelerate==1.5.2
aiohappyeyeballs==2.6.1
aiohttp==3.11.14
aiosignal==1.3.2
albucore==0.0.23
albumentations==1.4.24
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.9.0
astor==0.8.1
async-timeout==5.0.1
attrdict==2.0.1
attrs==25.3.0
babel==2.17.0
bce-python-sdk==0.9.29
beautifulsoup4==4.13.3
black==25.1.0
blinker==1.9.0
boto3==1.37.18
botocore==1.37.18
braceexpand==0.1.7
Brotli==1.1.0
cachetools==5.5.2
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
cloudpickle==3.1.1
coloredlogs==15.0.1
colorlog==6.9.0
contourpy==1.2.1
cryptography==44.0.2
cssselect==1.3.0
cssutils==2.11.1
cycler==0.12.1
Cython==3.0.12
datasets==3.4.1
decorator==5.2.1
# detectron2==0.6
dill==0.3.8
distro==1.9.0
doclayout_yolo==0.0.2b1
einops==0.8.1
et_xmlfile==2.0.0
evaluate==0.4.3
exceptiongroup==1.2.2
fairscale==0.4.13
fast-langdetect==0.2.5
fasttext-predict==0.9.2.4
filelock==3.18.0
fire==0.7.0
Flask==3.1.0
flask-babel==4.0.0
flatbuffers==25.2.10
fonttools==4.56.0
frozenlist==1.5.0
fsspec==2024.12.0
ftfy==6.3.1
future==1.0.0
fvcore==0.1.5.post20221221
grpcio==1.71.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
huggingface-hub==0.29.3
humanfriendly==10.0
hydra-core==1.3.2
idna==3.10
imageio==2.37.0
imgaug==0.4.0
iopath==0.1.9
itsdangerous==2.2.0
Jinja2==3.1.6
jiter==0.9.0
jmespath==1.0.1
joblib==1.4.2
kiwisolver==1.4.7
lazy_loader==0.4
lmdb==1.6.2
loguru==0.7.3
lxml==5.3.1
magic-pdf==1.2.2
Markdown==3.7
MarkupSafe==3.0.2
matplotlib==3.9.4
more-itertools==10.6.0
mpmath==1.3.0
multidict==6.2.0
multiprocess==0.70.16
mypy-extensions==1.0.0
networkx==3.2.1
numpy==1.26.4
omegaconf==2.3.0
onnxruntime==1.16.3
openai==1.68.2
opencv-contrib-python==4.6.0.66
opencv-python==4.6.0.66
opencv-python-headless==4.11.0.86
openpyxl==3.1.5
opt-einsum==3.3.0
packaging==24.2
paddleocr==2.7.3
paddlepaddle==3.0.0rc1
pandas==2.2.3
pathspec==0.12.1
pdf2docx==0.5.8
pdfminer.six==20231228
pillow==10.2.0
platformdirs==4.3.7
portalocker==3.1.1
premailer==3.10.0
propcache==0.3.0
protobuf==3.20.3
psutil==7.0.0
py-cpuinfo==9.0.0
pyarrow==19.0.1
pyclipper==1.3.0.post6
pycocotools==2.0.8
pycparser==2.22
pycryptodome==3.22.0
pydantic==2.10.6
pydantic_core==2.27.2
PyMuPDF==1.24.14
pyparsing==3.2.1
python-dateutil==2.9.0.post0
python-docx==1.1.2
python-dotenv==1.0.1
pytz==2025.1
PyYAML==6.0.2
rapid-table==1.0.3
RapidFuzz==3.12.2
rapidocr-onnxruntime==1.4.4
rapidocr-paddle==1.4.5
rarfile==4.2
regex==2024.11.6
requests==2.32.3
robust-downloader==0.0.2
s3transfer==0.11.4
safetensors==0.5.3
scikit-image==0.24.0
scikit-learn==1.6.1
scipy==1.13.1
seaborn==0.13.2
shapely==2.0.7
simsimd==6.2.1
six==1.17.0
sniffio==1.3.1
soupsieve==2.6
stringzilla==3.12.3
struct-eqtable==0.3.2
sympy==1.13.3
tabulate==0.9.0
tensorboard==2.19.0
tensorboard-data-server==0.7.2
termcolor==2.5.0
thop==0.1.1.post2209072238
threadpoolctl==3.6.0
tifffile==2024.8.30
timm==0.9.16
tokenizers==0.19.1
tomli==2.2.1
torch==2.3.1
torchvision==0.18.1
tqdm==4.67.1
transformers==4.42.4
typing_extensions==4.12.2
tzdata==2025.1
ultralytics==8.3.94
ultralytics-thop==2.0.14
unimernet==0.1.2
urllib3==1.26.18
visualdl==2.5.3
Wand==0.6.13
wcwidth==0.2.13
webdataset==0.2.100
Werkzeug==3.1.3
xxhash==3.5.0
yacs==0.1.8
yarl==1.18.3
boto3==1.26.0
